test(calldata): co-locate verifier tests
