+++ suite-common/calldata/src/verifier/evm/allowance.test.ts
@@ -1,6 +1,6 @@
-import { buildAllowance } from '../../../builder/evm/allowance';
-import { asEvmAddress } from '../../../types/evm';
-import { Verifier } from '../../../verifier';
+import { buildAllowance } from '../../builder/evm/allowance';
+import { asEvmAddress } from '../../types/evm';
+import { Verifier } from '../../verifier';
 
 const OWNER = asEvmAddress('0x9ea3721b5bf3b64b4418c38b603154d2d597fae3');
 const SPENDER = asEvmAddress('0x1231deb6f5749ef6ce6943a275a1d3e7486f4eae');
