+++ suite-common/calldata/src/verifier/evm/redeem.test.ts
@@ -1,49 +1,49 @@
 import { BigNumber } from '@trezor/utils';
 
-import { buildWithdraw } from '../../../builder/evm/withdraw';
-import { asEvmAddress } from '../../../types/evm';
-import { Verifier } from '../../../verifier';
+import { buildRedeem } from '../../builder/evm/redeem';
+import { asEvmAddress } from '../../types/evm';
+import { Verifier } from '../../verifier';
 
 const SENDER = asEvmAddress('0x9ea3721b5bf3b64b4418c38b603154d2d597fae3');
 const RECEIVER = SENDER;
 const OWNER = SENDER;
-const ASSETS = 1000000n;
+const SHARES = 1000000n;
 
-const withdrawHex = buildWithdraw(
+const redeemHex = buildRedeem(
     {
-        assets: new BigNumber(ASSETS.toString()),
+        shares: new BigNumber(SHARES.toString()),
         receiver: RECEIVER,
         owner: OWNER,
     },
     { sender: SENDER },
 ).data!;
 
-describe('verifyWithdraw', () => {
+describe('verifyRedeem', () => {
     it('returns isValid true on full match', () => {
         expect(
-            Verifier.evm.erc4626.withdraw(withdrawHex, {
-                assets: ASSETS,
+            Verifier.evm.erc4626.redeem(redeemHex, {
+                shares: SHARES,
                 receiver: RECEIVER,
                 owner: OWNER,
             }),
         ).toEqual({ isValid: true, issues: [] });
     });
 
-    it('returns VALUE_MISMATCH on full mismatch', () => {
+    it('returns VALUE_MISMATCH on shares mismatch', () => {
         expect(
-            Verifier.evm.erc4626.withdraw(withdrawHex, {
-                assets: 999n,
+            Verifier.evm.erc4626.redeem(redeemHex, {
+                shares: 999n,
                 receiver: RECEIVER,
                 owner: OWNER,
             }),
-        ).toEqual({ isValid: false, issues: [{ code: 'VALUE_MISMATCH', field: 'assets' }] });
+        ).toEqual({ isValid: false, issues: [{ code: 'VALUE_MISMATCH', field: 'shares' }] });
     });
 
     it('returns isValid true when partial checked fields match', () => {
         expect(
-            Verifier.evm.erc4626.withdraw(
-                withdrawHex,
-                { assets: 999n, receiver: RECEIVER, owner: OWNER },
+            Verifier.evm.erc4626.redeem(
+                redeemHex,
+                { shares: 999n, receiver: RECEIVER, owner: OWNER },
                 ['receiver', 'owner'],
             ),
         ).toEqual({ isValid: true, issues: [] });
